update printf
support channel
 - USCI_A0
 - USCI_A1

support option
 - USCI_Ax, CLOCK_1M,  BR_9600
 - (XXX) USCI_Ax, CLOCK_1M,  BR_115200
 - USCI_Ax, CLOCK_8M,  BR_9600
 - USCI_Ax, CLOCK_8M,  BR_115200
 - USCI_Ax, CLOCK_16M, BR_9600
 - USCI_Ax, CLOCK_16M, BR_115200

printf/
        modified:   main.c
        new file:   myclock.c
        new file:   myclock.h
        new file:   myprintf.c
        new file:   myprintf.h
