+++ printf/myclock.h
@@ -1,0 +1,11 @@
+
+#ifndef __MYCLOCK_H
+#define __MYCLOCK_H
+
+#define CLOCK_1M    0
+#define CLOCK_8M    1
+#define CLOCK_16M   2
+
+void clock_init(unsigned char clkType);
+
+#endif
